+++ tests/Main/Models/MultiScopedCluster.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Baum\Tests\Main\Models;
+
+class MultiScopedCluster extends Cluster
+{
+    protected $scoped = ['company_id', 'language'];
+}
