+++ tests/Main/Models/OrderedCluster.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Baum\Tests\Main\Models;
+
+class OrderedCluster extends Cluster
+{
+    protected $orderColumn = 'name';
+}
