+++ tests/Main/Models/ScopedCluster.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Baum\Tests\Main\Models;
+
+class ScopedCluster extends Cluster
+{
+    protected $scoped = ['company_id'];
+}
